+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased] Unreleased
 
+## Added
+
+- Support for PHPUnit 10 and 11.
+- The `MysqlServerController` extension.
+
+### Fixed
+
+- Ensure the `WPLoader` module will initialize correctly when used in `loadOnly` mode not using the `EventDispatcherBridge` extension.
+
 ## [4.2.5] 2024-06-26;
 
 ### Changed